+++ R/polmineR.R
@@ -19,6 +19,7 @@ NULL
 setOldClass("htmlwidget")
 
 
+
 #' polmineR-package
 #' 
 #' A library for corpus analysis using the Corpus Workbench (CWB) as an